Update export.class.php
Allow usage of a method param that is not part of the $obj from the sql request. 

The issue I encontered is the usage of order's LibStatut method that required a $mode param that can't be fetched from the sql request obj.

This would be the code for the export declaration : 
		$this->export_special_array[$r] = array(
			'orders.fk_statut'=> [
				'rule' => 'compute',
				'class' => 'Commande',
				'classfile' => '/commande/class/commande.class.php',
				'method' => 'LibStatut',
				'method_params' => ['orders_fk_statut', 'orders_facture', '1'],
			],
		);

We can see that the first two params will come from the $obj, but the third one will be used for it's value (in this case '1').